+++ lib/dart_fusion.dart
@@ -1,0 +1,22 @@
+/// A library that brings together a harmonious blend of essential tools, utilities, and components designed to supercharge your Dart projects.
+library dart_fusion;
+
+import 'package:equatable/equatable.dart';
+import 'package:flutter/foundation.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_svg/flutter_svg.dart';
+import 'package:go_router/go_router.dart';
+import 'package:logger/logger.dart';
+
+part 'src/d_behavior.dart';
+part 'src/d_image.dart';
+part 'src/d_log.dart';
+part 'src/d_state_widget.dart';
+part 'src/d_builder.dart';
+part 'src/d_extensions.dart';
+part 'src/d_button.dart';
+part 'src/d_models.dart';
+part 'src/d_overlay.dart';
+part 'src/d_typedefs.dart';
+part 'src/d_parse.dart';
+part 'src/d_textarea.dart';