+++ lib/src/d_typedefs.dart
@@ -1,0 +1,4 @@
+part of '../dart_fusion.dart';
+
+/// Define [JSON] Type as a shortcut for [Map<String, dynamic>].
+typedef JSON = Map<String, dynamic>;